refactor: extract normalizeDNSName to idna package for reuse
