+++ internal/idna/idna.go
@@ -17,6 +17,9 @@ limitations under the License.
 package idna
 
 import (
+	"strings"
+
+	log "github.com/sirupsen/logrus"
 	"golang.org/x/net/idna"
 )
 
@@ -27,3 +30,16 @@ var (
 		idna.StrictDomainName(false),
 	)
 )
+
+// NormalizeDNSName converts a DNS name to a canonical form, so that we can use string equality
+// it: removes space, get ASCII version of dnsName complient with Section 5 of RFC 5891, ensures there is a trailing dot
+func NormalizeDNSName(dnsName string) string {
+	s, err := Profile.ToASCII(strings.TrimSpace(dnsName))
+	if err != nil {
+		log.Warnf(`Got error while parsing DNSName %s: %v`, dnsName, err)
+	}
+	if !strings.HasSuffix(s, ".") {
+		s += "."
+	}
+	return s
+}
