+++ plan/plan.go
@@ -19,7 +19,6 @@ package plan
 import (
 	"fmt"
 	"slices"
-	"strings"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
@@ -142,7 +141,7 @@ func (t *planTable) addCandidate(e *endpoint.Endpoint) {
 
 func (t *planTable) newPlanKey(e *endpoint.Endpoint) planKey {
 	key := planKey{
-		dnsName:       normalizeDNSName(e.DNSName),
+		dnsName:       idna.NormalizeDNSName(e.DNSName),
 		setIdentifier: e.SetIdentifier,
 	}
 
@@ -371,19 +370,6 @@ func filterRecordsForPlan(records []*endpoint.Endpoint, domainFilter endpoint.Ma
 	}
 
 	return filtered
-}
-
-// normalizeDNSName converts a DNS name to a canonical form, so that we can use string equality
-// it: removes space, get ASCII version of dnsName complient with Section 5 of RFC 5891, ensures there is a trailing dot
-func normalizeDNSName(dnsName string) string {
-	s, err := idna.Profile.ToASCII(strings.TrimSpace(dnsName))
-	if err != nil {
-		log.Warnf(`Got error while parsing DNSName %s: %v`, dnsName, err)
-	}
-	if !strings.HasSuffix(s, ".") {
-		s += "."
-	}
-	return s
 }
 
 func IsManagedRecord(record string, managedRecords, excludeRecords []string) bool {
