+++ plan/plan_test.go
@@ -1054,104 +1054,6 @@ func validateEntries(t *testing.T, entries, expected []*endpoint.Endpoint) {
 	}
 }
 
-func TestNormalizeDNSName(tt *testing.T) {
-	records := []struct {
-		dnsName string
-		expect  string
-	}{
-		{
-			"3AAAA.FOO.BAR.COM    ",
-			"3aaaa.foo.bar.com.",
-		},
-		{
-			"   example.foo.com.",
-			"example.foo.com.",
-		},
-		{
-			"example123.foo.com ",
-			"example123.foo.com.",
-		},
-		{
-			"foo",
-			"foo.",
-		},
-		{
-			"123foo.bar",
-			"123foo.bar.",
-		},
-		{
-			"foo.com",
-			"foo.com.",
-		},
-		{
-			"foo.com.",
-			"foo.com.",
-		},
-		{
-			"_foo.com.",
-			"_foo.com.",
-		},
-		{
-			"\u005Ffoo.com.",
-			"_foo.com.",
-		},
-		{
-			".foo.com.",
-			".foo.com.",
-		},
-		{
-			"foo123.COM",
-			"foo123.com.",
-		},
-		{
-			"my-exaMple3.FOO.BAR.COM",
-			"my-example3.foo.bar.com.",
-		},
-		{
-			"   my-example1214.FOO-1235.BAR-foo.COM   ",
-			"my-example1214.foo-1235.bar-foo.com.",
-		},
-		{
-			"my-example-my-example-1214.FOO-1235.BAR-foo.COM",
-			"my-example-my-example-1214.foo-1235.bar-foo.com.",
-		},
-		{
-			"點看.org.",
-			"xn--c1yn36f.org.",
-		},
-		{
-			"nordic-ø.xn--kitty-點看pd34d.com",
-			"xn--nordic--w1a.xn--xn--kitty-pd34d-hn01b3542b.com.",
-		},
-		{
-			"nordic-ø.kitty😸.com.",
-			"xn--nordic--w1a.xn--kitty-pd34d.com.",
-		},
-		{
-			"  nordic-ø.kitty😸.COM",
-			"xn--nordic--w1a.xn--kitty-pd34d.com.",
-		},
-		{
-			"xn--nordic--w1a.kitty😸.com.",
-			"xn--nordic--w1a.xn--kitty-pd34d.com.",
-		},
-		{
-			"*.example.com.",
-			"*.example.com.",
-		},
-		{
-			"*.example.com",
-			"*.example.com.",
-		},
-	}
-	for _, r := range records {
-		tt.Run(r.dnsName, func(t *testing.T) {
-			gotName := normalizeDNSName(r.dnsName)
-			assert.Equal(t, r.expect, gotName)
-		})
-	}
-}
-
 func TestShouldUpdateProviderSpecific(tt *testing.T) {
 	for _, test := range []struct {
 		name         string
